Feedback from code review. More checks.
