+++ messages/test-ui-run-mobile.json
@@ -1,0 +1,26 @@
+{
+    "commandDescription": "Run a UTAM test.",
+    "platformFlagDescription": "Specify platform (iOS or Android).",
+    "configFlagDescription": "Specify a path to WDIO config file with UTAM configuration.",
+    "specFlagDescription": "Specify a path to test spec.",
+    "runningUtamTest": "Running UTAM test",
+    "reqs:wdio:title" : "Checking if WDIO is installed",
+    "reqs:wdio:fulfilledMessage" : "A valid version of WDIO is installed: %s.",
+    "reqs:wdio:unfulfilledMessage": "Couldn't find WDIO installed: %s.",
+    "reqs:appium:title" : "Checking if Appium is installed.",
+    "reqs:appium:fulfilledMessage" : "A valid version of Appium is installed: %s.",
+    "reqs:appium:unfulfilledMessage" : "Couldn't find Appium installed: %s",
+    "reqs:appiumDriver:title" : "Checking if Appium driver is installed.",
+    "reqs:appiumDriver:fulfilledMessage" : "Appium driver installed: %s.",
+    "reqs:appiumDriver:unfulfilledMessage" : "Couldn't find Appium driver installed: %s.",
+    "reqs:appiumService:title" : "Checking if @wdio/appium-service is installed.",
+    "reqs:appiumService:fulfilledMessage" : "@wdio/appium-service is installed.",
+    "reqs:appiumService:unfulfilledMessage" : "@wdio/appium-service is not installed.",
+    "error:invalidPlatformFlagsDescription": "Invalid input value for platform.",
+    "error:invalidConfigFlagDescription": "Invalid or missing value for WDIO configuration.",
+    "error:invalidSpecFlagDescription": "Invalid or missing value for test spec.",
+    "error:unexpectedErrorDescription": "Unexpected error: %s",
+    "error:configFileDoesntExistDescription": "Config file doesn't exist.",
+    "error:specFileDoesntExistDescription": "Spec file doesn't exist.",
+    "error:wdioIsntInstalledDescription": "Make sure that wdio is installed."
+}
