--- messages/test-utam-run.json
@@ -1,15 +1,0 @@
-{
-    "commandDescription": "Run a UTAM test.",
-    "platformFlagDescription": "Specify platform (iOS or Android).",
-    "configFlagDescription": "Specify a path to WDIO config file with UTAM configuration.",
-    "specFlagDescription": "Specify a path to test spec.",
-    "runningUtamTest": "Running UTAM test",
-    "checkWdio": "Check if WDIO is installed",
-    "error:invalidPlatformFlagsDescription": "Invalid input value for platform.",
-    "error:invalidConfigFlagDescription": "Invalid or missing value for WDIO configuration.",
-    "error:invalidSpecFlagDescription": "Invalid or missing value for test spec.",
-    "error:unexpectedErrorDescription": "Unexpected error: %s",
-    "error:configFileDoesntExistDescription": "Config file doesn't exist.",
-    "error:specFileDoesntExistDescription": "Spec file doesn't exist.",
-    "error:wdioIsntInstalledDescription": "Make sure that wdio is installed."
-}
