+++ package.json
@@ -64,7 +64,7 @@
     "lint-staged": "^13.1.0",
     "oclif": "^3.4.6",
     "prettier": "^2.8.3",
-    "salesforce-pageobjects": "^2.1.0",
+    "salesforce-pageobjects": "3.0.1",
     "ts-jest": "^29.0.5",
     "typescript": "^4.9.4",
     "wdio-chromedriver-service": "^7.3.2",
